tidy

+++ gfx/ui/vertical_layout.cpp
@@ -39,7 +39,7 @@ void VerticalLayout::add_widget(std::unique_ptr<Widget>&& widget)
         focus_ = children_.front().widget.get();
     for (size_t i = 0; i < children_.size(); ++i) {
         float size = 2.0f / children_.size();
-        children_[i].transform = {{0.0f, size / 2.0f + (children_.size() - 1 - i) * size - 1.0f},
+        children_[i].transform = {{0.0f, size / 2.0f + static_cast<float>(children_.size() - 1 - i) * size - 1.0f},
                                   {1.0f, size / 2.0f}};
     }
 }
